Fix R backend
The source of the madness was the line

```
{par}  { yylval = yytext; ... }
```

In the R.l lexer. `yytext` is a pointer to a character vector used
internally in flex. `yylval` is the semantic value of the acquired
token. By setting yylval equal to yytext, I made the token semantic
value unstable. The solution is to copy the value:

```
{par}  { yylval = strdup(yytext); ... }
```

I had made the same mistake in the frontend lexer. This had led to no
end of madness, as things I thought ought to work did not, and changes I
thought should not fix any problems, did.