Implemented a nearly complete parser
Three features are missing:

1) comma lists, e.g.

```
@check
A,B,C :: true
```

which is syntactic sugar for

```
@check
A :: true
B :: true
C :: true
```

2) conditionals

```
A --> (V ? B1 : B2) --> C
```

In addition, the bison reports the warning:

```
rat.y: warning: 3 shift/reduce conflicts [-Wconflicts-sr]
```

3) group handling

Similar to 1), groups, such as the builtin __all__ should automatically
split.

While this parser is nearly complete, in that it touches nearly all the
features I want for the frontend, I am not happy with how everything
works. Specifically arrays and loop handling is (e.g. LOOP_SPLIT_ON) is
ugly.

On a deeper level, the distinction between function, manifold emitter,
and manifold instance is weak. I treat them all with the same name,
leaving the backends to sort the mess. Which they can't. I don't give
them enough information. I need to create a unique label for each
manifold that is emitted. The LABEL token allows overriding of the
automatic label. Using the name of an emitter without a label, implies
the action is performed on all instances.

I might also add a @require section, where all functions that must be
present in the environment are listed. This will require some extra
boilerplate on the users side. Maybe I could even add a @type section.
Then I could type check at compile time on the frontend side. Sets of
functions and their types could be imported. All the functions in the
execution space could be written to a rat header and imported by any
workflows wishing to work in that space. The spaces would want to be
able to do this anyway, since it is needed for searching. This is part
of their role as function databases. Of course, @require and @type
sections could be merged to one @declaration section, with haskell style
function declarations.

Maybe also a @version section? @dependency section, environment
validators? Is that even necessary? Individual manifolds can call
arbitrary validators on their own. These validators can see the
environment, the hardware, check versions, or anything they like.