Add support for conditionals
The term conditional isn't quite accurate. Given a path like

```
@type
x :: a -> b
V :: b -> c
z :: c -> d
A :: ?
B :: ?
C :: ?
@path
w --> x --> (V ? A, B, C) --> z
```

here is what happens:

 1. z requests data of type c from V
 2. V requests data of type b from x
 3. V does something and returns type c data

That is about as specific as I can be, everything else is dependent on
the implementation of V. If A, B, and C are all of type `b -> c`, V
might call one of them based, all of them and choose the one that gives
the highest value, randomly select one of them, or anything else. But A,
B, and C could be otherwise typed. Anything is fine, so long as V
ensures the ultimate output is of type c.

V is a manifold that takes manifold arguments in addition to the input
manifold.

Is this even needed? A, B, and C could just be expressed as standard
inputs to V. (where A B C don't call w)

```
(w --> x) A B C --> V --> z
```

I do need to add a distributive property to the paths, e.g. a shorthand
for

```
(w --> A) (w --> B) (w --> C) --> V
```

maybe `w --> A B C --> V`