Add manifold structure
So far I have been immediately printing a RLI command for each elemental
command in that rat file.

However, the main function of rat is to build a set of manifolds, so it
makes more sense to build a list of manifolds. After the parse is
complete, I will have a list of manifolds. I can iterate through it,
checking for errors, asserting type correctness, etc. Then, if all is
well, I can print the RLI output.

I don't need to build an AST; a list of manifolds will be cleaner.

In this commit, I have written the manifold structure, but have not yet
connected it to the parser.