Allow positionals to be placed in path
Before, all variables in a path where identical with their output. Any
constants were given to the manifolds as arguments in the @arg section.
The idea was that if it is not incoming data, and is constant, it isn't
interesting, and should not clutter the @path section. However, this
does not fairly represent control functions and logically required
constants, e.g.

```
@path
map . well
@arg
map :: function=square
```

The above gets a list of data and maps the function square over it.
@path leaves out the most important piece of the workflow, `square`.
With the new syntax, I can say

```
@path
map . &square well
```

Another case where a constant is logically important is specifying the
dimension over which mapping will occur , e.g.

```
@type
apply :: [[a]] -> Integer -> ([a] -> b) -> [b]
@path
apply . well &1 &sum
```

Where the data can be mapped row-wise (1) or column-wise (2).

`sum` here will be treated as a primitive function, not a manifold, it
cannot call effects or by validated. It has to be a primitive function
because it is being passed into native function where it will be passed
arguments.

Principle: @path should be a sufficient description of the workflow.
@type gives context and character; @arg fills in details; @check
specifies invariants and dependencies; @effect produces the output;
@cache, @pass, @fail, @pack, @unpack are all boilerplate.