Add a language-specific source section
Rat takes takes a general pipeline and converts it into a specific
language. This will usually require importing a function space in the
destination code. There may be other setup to do as well. Or you might
want to write a wrapper for a function in language X right there in the
Rat code. So I have added a section for arbitrary code entry. This code
will be pasted verbatim to the beginning of the output file.

Currently, the vim-syntax file will load the correct highlighting for R
code, and I can easily extend it to any other language. However, I
should automate it, rather than hard the supported languages.

I am also submitting a sample file `small.rat` that sources R code.

None of this is yet supported by the lexer or parser.