Resolved all Bison conflicts and warnings.
Known issues:

 * ontology section does not allow space separated constructors, e.g.
   `Tree :: Leaf Value | Node Tree Tree`
 * type section does not support parentheses
 * path section does not support multiple inputs, e.g. `a . b c`

Several problems are arising from my insistence on not requiring end of
statement characters (e.g. ;), not conditioning on whitespace, yet using
whitespace as an operator in lists (e.g. `a . b c`).

I want to write this:

```
@ontology
a :: b c | d e
f :: g
```

This is a problem for the next commit