Wired paths (but ignore /.*)
In the previous commit I made the Label stucture and built it into the
lexer. Here I change the type of IDENTIFIER to `Label*`. I have handled
all side-effects of this change. Everything compiles and runs exactly as
it did before.

I changed the `NamedList->name` type from `char*` to `Label*`. This is
reasonable for the sections that alter manifolds. But it is a little
crufty for the other sections, where `name` will be a linked list
of constant length 1.

NOTE: I have not extended this change to output.