Rename from Rat to Loc
The name Rat was a vague reference to my R pipeline package Pied, which
was in turn a reference to my favorite musician. Rat was initially an
auxillary program for Pied, however now the situation is reversed.

I do not value acronyms -- yet one so succinct and descriptive I can
hardly reject: Language of Composition. The name itself may ward against
contaminants and aid explanations.