Add group references to lexer
It still breaks in the parser. The group ref is just a character string.
After the symbol table is complete, and all paths have been recorded,
these grprefs need to be expanded. For this I need a table_clone
function. Then I need to extend many of the table.c functions to
understand the new member.

So much to do ...