Add stub support for manifold arguments
Passing manifolds, nests of manifolds, groups, whatever else as
arguments is a grand idea. However implementation is a bit tricky.
Manifolds do not take user input. The calls to input manifolds is
enclosed in the function. To pass the manifolds, we need to "open" them.
We also need to figure out the type (what inputs can it take?). We need
to count the outputs. Probably, there should be only one output.

Types are fun. I think once I add back the Haskell style type section,
it will be pretty straightforward to infer and check types. I already
have a method for implicit type linking (in build.c). With just a bit of
love, I'll have type inference for any given path. I can reduce them to
a single function ... a bit tricky. But most of the difficulty is
probably on the backend. Ahhh, the backend. The dreaded backend. The
neglected backend ...