Add untested skeleton of new infrastructure
I need lists. Many different kinds of lists. The overall symbol table is
a heterogenous list. The compositions are a heterogenous list of limited
types. The left-hand-side paths are lists of strings, possibly nested
lists of strings.

Previously I have used Table and Entry for most of the structure. Then
ad hoc lists for Path and Selection. Entry was used in places with no
lhs, where it would be more intuitive to have a simpler list.

I've now added a skeleton of a new system. The structure W holds a
class, a link to the next W, and a value. The value is a union of all
possible types. The class is an enum of all supported classes. Jump
tables will be used to control handling of the values.

I also added prototypes for list functions, including some higher order
functions.

The W class will wrap entry and will be used by Table. It will subsume
the list handling of Path and Selection. Eventually it will be used to
handle to multiple inputs of the couplets.

I rename the Entry strucutre 'Couplet', since it is the union of a lhs
and rhs.