Distinguish C_MANIFOLD and P_MANIFOLD
The variables contained in a path represent manifolds. But are not
identical to the actual manifolds. That is, the `x` in `A :: x . y` is a
label associated with the actual manifold. In other words, it is a
couplet, joining a label and a true manifold. But a Couplet structure is
a set of two W things, so I need to wrap the true manifold in a W
container, hence the need for P_MANIFOLD (where P means 'primitive').