Implemented export section
There are two expressions that are legal in the export section:

 1. A
 2. B as C

Where A, B and C are variable names (following same syntax rule as other
variables).

These compile into the LIL code

```
EXPORT A A
EXPORT B C
```

That is, if no alias is specified with `as`, the alias defaults to the
original name.

I currently don't do any checking on whether such an export is legal.
Such a check would require specifying which things can be exported
(manifolds, path names, etc). There are a lot of technicalities to
consider. I had mostly thought of this as a way of producing a few
(usually one) means of invoking the pipeline, a little like declaring a
main.

This could also be important in libraries, where only the top level
compositions should be available. Although currently my import section
just reads in the whole file.

So export is all rather formless and void at the moment. But hey! If
exporting makes you happy, export away!