Add alias section
The aliases currently do exactly one thing: replace the function name of
the specified manifold(s). They are parsed into Couplet's, exactly like
all the other modifiers, but rather than appending a value to a list of
manifold modifiers, they replace the name of the function.

One of the ultimate uses for alias will be to tie general operations to
language specific function names. So eventually I will need to elaborate
the current system. But my first goal is to build a working LOC system
for a single language.