Implemented source code section
I toyed with various ideas about how to delineate the source regions in
LIL. My decision: indentation. E.g.

```
@source R
x + 1
foo
```

Goes to

```
SOURCE R
    x + 1
    foo
```

This is should be pretty easy to parse and there is no risk of my
sentinal value conflicting with the code. I break on newlines, so the
LOC code has no way to circumvent the padding.