Add type section parsing
Type declarations can now be parsed. Nesting is legal, but list syntax
is not.

All these are legal

```
@type
foo :: a
bar :: a -> b
baz :: a -> (a -> b) -> b
```

But I can't currently do, for example

```
@type
foo :: a -> [b]
```

Allowing operations of this sort would be another possible path to
looping.

I also don't allow type constructors.

All that fancy stuff I'll save for later. The next goal is simply to
check that the correct number of arguments are being passed to each
manifold.

Then write a function to compose a path into a single function and infer
it's type. This will be useful for the deref operations.