Cleaned up sample files
I moving toward a test suite. My approach is to write simple scripts and
their expected output, then check the diffs. It isn't perfect; I would
prefer unit tests on my functions; but it suffices to control
regression.

There is a lot more work to do in developing minimal test cases, but
this is the start.