Release v0.2.0
v0.2.0 includes most of the material I need for building compositions of
manifolds, importing files, and setting/checking function types. It
lacks good support for loops and conditionals (you could hack it). That
will probably be the focus of v0.3.0. I also have some very theoretical
ideas about writing plum pudding functions rather than the implicit and
singular manifold (v0.4?).

Note: In a previous commit I mentioned the coming release of 0.1.0. But
I actually already released that and current code in locr is dependent
on it.

Note: Also the tests and error handling I promised are still not ready.