Add language section
LOC is a meta-programming language that provides a structure for
external functions. In a monoglot workflow, one can simply use a
language specific backend to compile the LOC into a single-language
output. However, LOC should allow seamless, boilerplate-free integration
of any combination of languages. With this commit the language of a
manifold may be specified.

```
@path
A :: g . f
@lang
g :: R
```

compiles to
```
emit R g
emit * f
```

The `*` implies the manifold is universal, and may be assumed to be
defined in any language.

When a polyglot workflow is compiled, a single LIL is produced and
compiled into an executable for each language and a master controller
(probably a shell script). The language executables take a manifold id
as an argument and then will run it and cache the value. When a manifold
in a foreign language is encountered (a trans-manifold), a wrapper
manifold is created in its place which will call the foreign language
executable and then call itself. Cache frameworks should be universal
(e.g. nocache, filecache, sqldb) with keys based on the manifold id. A
manifold cannot take input from a foreign manifold with a non-persistent
cache.