Full working backend for Bash
For each language in the LOC file, I will create an executable program
called `call-LANG`. These executables take a single argument: a manifold
id. They return the manifold output.

The first step is to compile the LOC code into LIL. Then for each
language 'LANG' I

 1. extract any source code in LIL that is written LANG

 2. generate all manifolds in native LANG

 3. concatenate the base library (`lib/LANG-base.loc`), the source code
 from step 1, the manifolds from step 2, and language specific
 boilerplate needed to take an argument from the command line and call
 the appropriate manifold. If the manifold does not exist, it should
 return the appropriate error message.

Once these are all complete, I can build the master executable, which
exposes the exported functions to the user.

If a given input to a manifold is from a manifold in the same language,
I call this a cis-input. If the input is from a foreign language, it is
a trans-input. cis-inputs are simple function calls, requiring no
special handling. trans-inputs call the foreign executable and collect
the output as STDOUT. This may require special `open` functions to parse
the raw text into a native structure.

A better name for the language-specific executables is, perhaps,
manifold pools. trans-calls have to leave the native-pool and access a
foreign one, this requires a universal data language, which of course is
simple text.

This also means that I need functions for dumping any native structure
into text and vise versa. I might call these show and read functions,
ala Haskell.

The open functions that access foreign pools will probably involve a
foreign `show` manifold that takes the foreign manifold of interest and
textifies it.

Hmm, native and universal representations. There will be a lot of wiring
to do here. Perhaps my commitment to parameter free manifolds is
misguided. Perhaps I should add a boolean argument that specifies
whether to return a universal or native value.

No, I can let the cachers hash this out. I could pass the universal
versus native flag in the Mb, the output container. Then the open
functions could handle the issue.