Add make install
The `make install` command now creates a `.loc` folder in the home
directory with the following contents:

    .
    ├── bin
    │   ├── loc
    │   └── parse.awk
    ├── etc
    │   └── grammar.m4
    └── lib
        └── bash-base.loc

The `build.sh` is copied to `$HOME/bin` with the name `loc`. Required
files and scripts are accessed by hardcoded paths to the `.loc` folder.

This allows LOC projects to be compiled anywhere.

Libraries can be imported into the `.loc/lib` folder and eventually I'll
make a locrc file (although, at the moment I'm not sure what would be
configurable).

Eventually I should also add texinfo pages and all that jazz. Yay,
documentation ...