Add hooks
I want to distinguish between effects and hooks. An effect is a function
of the manifold output. A hook is a manifold that is called after the
result is cached.

I am not at all happy with these words. I must find better ones.

Anyway, I have added support for hooks and effects to both the front and
backends.