Add manifold nexus
Now a manifold nexus is created with each build. This shell script
contains documentation for the manifolds and allows them to be called by
name (rather than id).

There is much work to be done here:

 1. the manifold nexus exports all functions, whereas it should export
 only the ones specified in the export section of LOC

 2. the frontend should resolve exported function to the root manifold.

 3. docstrings need to be incorporated into the help

 4. could add `remove`, `archive`, `list` and some other utilities to
 the nexus.

The workflow folder should also store the original loc file and the
nexus should provide access to it (and LIL).

Well, there is all sorts of stuff it could do ...

Maybe allow it to call the LOC typechecker.

Perhaps the workflow folder should be created in a `tmp` folder in the
`$HOME/.loc` directory. Then the manifold nexus can be created in the
users working directory. This way, the user gets one executable to work
with after compiling their LOC script.

It could also have functions for plotting the manifold network.