Allow vector types
I only allow one dimensional vector types, e.g. `[Number]`. This
involves only the addition of `\[{sym}\]' to the lexer. That is, there
is no higher-level handling.

I ought to allow lists of arbitrary depth and lists that contain
functions. I don't do this simply because the parsing is more
complicated, but I will certainly add it later.