Partial support for R
I still need to write the EPILOGUE and PROLOGUE, but the base manifolds
being built correctly.

I also changed the way manifold pools are named. Rather than
`call-bash.sh`, I changed it to just `call.sh`. Now the extensions are
the sole language determinant. The extensions are also the LANG strings
that should be used in LOC.