Add C_REFER token for manifold reference
The syntax is as follows:

```
@path
Well :: runif
Max :: max . <runif>
```

The Max path refers to `Well/runif`. In contrast, the two runif's below
are completely different instances of the runif manifold and will be
cached separately.

```
@path
Well :: runif
Max :: max . runif
```

The tokens are currently passed, but I still need to wire C_REFER to
inputs.