Bux fixes and REFER test
The R test works fine. However, I encountered a very interesting bug on
the Bash side. It occurs in the following compiled code:

```
random () {
    echo $RANDOM
}
m0(){
    if datcache_chk m0
    then
        datcache_get m0
    else
        random | datcache_put m0
    fi
}
m1(){
    cat  <(m0)  <(m0)
}
```

The first time this is run it prints two unequal lines, e.g.

```
18346
3984
```

The next time it is called, it produces the expected output, e.g.

```
18346
18346
```

What is happening here is a race condition. The two m0's are called in
m1 before anything is cached. One of the two will,
non-deterministically, be cached first.

This is a deep problem. I do yet not have an elegant solution.

I will leave this test case to fail painfully with every run.