Extend tests and fix imports
Previously, when multiple imports where included in one import section,
only the first was considered. This was because the first import
statement was basically replaced with the code of the file it
referenced. Since this file probably didn't end in an import statement,
the second import string is interpteted in the wrong context.

I fixed this by simply added a `@import` section header to the end of
each file.