Add full support for modify operators
Some of the behavior is a bit surprising. For example

```
@effect
a := foo
a :- foo
```

still results in an effect `foo` being attached to `a`. The reason is
that the two foo are distinguishable, they compile to different
manifolds with different uid. I determine identity of manifolds by uid,
not function name. For the above to work, you have to reference the
first foo.

```
@effect
a := foo
a :- <foo>
```

Also, I do not support the ':-' operator for arguments. For arguments,
checking of identity is not as straightforward. Do I just compare the
left-hand sides? What if there is no left-hand side? I can compare
right-hand side by identify. That works for things like unbundled Bash
flags. But what about bundled ones? What about positional arguments?
When is a "positional" really positional? The ':-' operator will
probably never work for arguments. However, ':+' and ':=' are perfectly
fine.