Add left-handed stars
Stars can now be used to match anything. This is especially useful for
languages, having to write

```
@lang
foo,bar,baz,print_this_giant_thing :: R
```

is pretty dang annoying. Now we can write:

```
@lang
* :: R
```

Yet this introduces two problems.

First, the `*` matches everything, even in files that are included. This
brings us back to the `include` versus `import` problem.

Secondly, everything is assigned to the *exact* thing on the right. This
is actually not a problem unique to `*`, it was present in lists as
well.

```
@path
A :: c . b . c
@effect
* := foo      # s1
a,b,c := bar  # s2
```

s1 and s2 both set a, b, and c to match an exact manifold (same uid).
This is equivalent to

```
@effect
a := foo
b := <foo>
c := <foo>
```

Maybe we can add the operators `:++` and `:--` operator (or something)
that give unique id to each foo, so `* :++ foo` is equivalent to

```
@effect
a := foo
b := foo
c := foo
```

Or we could add some magic to the rhs, e.g. `* :+ ^foo`, where `^foo`
indicates foo is recreated each time it is accessed. In which case it
would make sense to treat foo as an array, with element access `foo[a]`.

How about a rhs handle to the star values?

```
@effect
* :: print . <$$>
```

which would be equivalent to (for manifolds in {a,b,c})

```
@effect
a :: print . <a>
b :: print . <b>
c :: print . <c>
```

Well, this is all adorable, but I'll leave such magic and sugar later.

A perk of strong separation between front and back is that there can be
a million LOC's with a million syntactic flavors all compiling to the
same LIL. LOC is nimble. LIL, however, must be eternally unchanging.
Absolute immutable perfection. Changes to LIL may require changing all
the language-specific backends. Well, this isn't entirely true, since
the direct interface will LIL is the single lil-parser (in AWK
currently).

Even more immutable than LIL, is the general functional form of the
manifold. This is where the language-specific backends really differ.

I digress ...