Force evaluation in R null function
The R null function is designed to be a terminal point for a pipeline
that dumps inputs, so nothing is printed to STDOUT. However, my previous
implementation

```
null <- function(...) { NULL }
```

had a serious problem. It did not force evaluation of inputs. In context
a terminal null compiles to

```
m0 <- function(){
    Mb <- null ( m1() )
    ...
}
```

If the pipeline is called by calling `null` (e.g. `manifold-nexus.sh
null`), then null will be called and NULL returned, but since the
arguments are not used, they will not be evaluated (since R is lazy).

To fix this, I force evaluation:

```
null <- function(...) { list(...); NULL }
```

This fixes the 3rd known problem, `r-self-reference`.