Add test for open-manifold-based loop
The idea (not yet implemented) is to take a path and 'open' it, so it
takes arguments. These arguments are given by the calling function, and
are passed through all the manifolds until they meet someone who
actually wants them.

Building the open-manifolds will require a lot of work on both ends. A
real pain. But it will totally rock if I can get it working. The
open-manifolds will have all the flexibility of a standard manifolds,
all the effects checks and all that. I'll need to write in counters of
some sort, to distinguish between the isomers (specifically in the
caches). There will be no way, now or ever, to distinguish between them
in LOC. Only their inputs will distinguish them.

I could add a universal counter. Something that is incremented every
time anything accesses it. This could be used to generate ids across the
program. The approach would break down if work was being exported to an
outside cluster, but this could be rememdied with composite ids.