Map functions are working in R
However, modifications of open manifolds are not possible. The reason is
that the modification manifolds (hook, for example) also need to be
open.

So I need to recurse into all manifold modifiers when I open a manifold.
I can do that.

Here is my R loop example

```
map  .
    & (plot_pdf . hclust $1 . dist . normalize . load_data . $1)
      (c . "iris" "cars" "mtcars")
```

The function 'c' in are is used to contatenate elements into a vector.
'map' takes takes a list of three strings and applies each to the
anonymous composition '& (...)'. The LOC compiler infers the number of
positional arguments that should be passed into the composition by
finding the larget '$n', in this case 1.

Note I have the plot_pdf function in the main path, rather than as a
hook. If I use plot_pdf as a hook, it will try to call the hclust
manifold with no arguments, which fails. On a deeper level, plot_pdf is
outside the superposition, it can't distinguish between the many
hclusts.