Add (somewhat buggy) R support for open manifolds
This is still a little icky ... for example, a modifier can't access the
first element in a deref. It treats it as a naked function, maybe this
is fine. I can get around the problem by just adding a null function as
the first element. Of course this only works because I run the loop for
effects only. I think I could replace the null with id. Anyway, I will
need to experiment and test.

Also, I've broken a bunch of other things in the process. I will need to
fix them next.