Allow -----+ as separator (does nothing)
This is purely a convenience, probably an unnecessary one, since
comments have the same effect. But now you can do

```
@path
asdf dfdf .
    sdfdf .
    dfdf
--------------------------------------
A :: asdfdfdf . dfdfdfdf . sdfd
```

Pretty pointless ... but it looks nice? Um, and it is highlighted so
prettily in vim. It is kind of a reanimated historic artifact from when
I used it as a statement separator.