Add map to R and sh core
There are certain control functions that every language should have. And
I want LOC to have functions that work exactly the same in every
language. For now, I am just adding a map function. It takes a function
and a list of inputs (in that order) and applies the function to each
element in the input, returning a list of outputs.

I am accreting far too many little core libraries. I need to consolidate
in some way. Maybe also automatically include the core functions.