Thoughts on manifolds
A manifold has the form

```
m1 <- function (){
    # @0
    if(cache_chk("m1")){
        # @8
        b = cache_get("m1")
        # @9
    }
    else{
        # @2
        if( [check] ){
            # @4
            b = fun([input], [arg])
            cache_put("m1", b)
            # @5
        } else {
            # @6
            b = fail()
            cache_put("m1", b)
            # @7
        }
        # @3
    }
    # @1
    return b
}
```

Where `# @i` is an effect. The effects are manifolds, thus take no
arguments. If they produce output, it is disposed of quietly by the
calling manifold. To this point, I have distinguished between effects
and hooks (`@5` and `@3` above). But they are fundamentally the same.
Furthermore, they are members of a larger, marginalized community
(`@0-9`). There are enough possible places for an effect hook, that I
hesitate to name them. Do I really want section titles like
`@precache_prehook`, `@precheck_posthook`, etc? Shall I be verbose or
arcane? I choose arcane, for now at least. Again, new flavors or LOC can
easily be added.

***

I also briefly considered wrapping the execution of `fun` in a try-catch
block. But I decided against it. We need explicit exception handling, we
can use compositions:

```
  @path
  bar . (tryCatch . foo) . ladida
```

Of course you could also pull out the nests, maybe you want to wrap
everything in tryCatch:

```
  @path
  *Bar . *Foo . ladida
  Foo :: tryCatch . foo
  Bar :: tryCatch . bar
```

A problem with this is that we are wrapping the manifold call, not the
function call. We could deref the manifolds:

```
  @path
  bar . *Foo . ladida
  Foo :: (tryCatch . &foo $1)
```

Maybe we simplify this with exceptional sugar, but I'll save that for
later.

***

I am also planning on removing the `pack`, `open`, and `run` modifiers.
All of these can be replicated with compositions. Perhaps I will bring
them back in the future. But for now I need a minimal, reliable, simple
system.

PS. Here I commit only notes. The reimplementation will follow.