10 failing
There is a necessity to pipe output through more to allow it to be
written to a file in certain contexts. I've seen this before in some
Bash loops. Some sort of fraying of outputs? Really not sure. Output is
printed, but I cannot directly redirect it to a file (either as STDOUT
or STDERR).