Add minimal and borken type handling
Now all LOC scripts have to import a type library. The default one is
`core/type.loc`. It should contain everything that is needed to handle
typed foreign calls, type conversions, type printing, and type
validation.

Universal representation can be either a literal string or a filename
referencing data in a universal form. The `universal_to_native` function
converts this to a native structure based on the expected type. For
example, 'Int' is a literal. `native_to_universal` performs the reverse.

The manifold nexus prints the universal representation to STDOUT and
STDERR (if present). It expands files.

Problems:
  * r-logical test spins into an infinite loop
  * half the tests fail, including all tests in type section
  * type handling is basically limited to Int