20 fail - fix handling for NULL outputs
How to deal with STDOUT? It is convenient to be able to hang functions
where ever I like and have messages printed out. However, this is highly
effectual. It can easily interfere with other parts of the program. It
is better to write to files or to stderr. But this requires some extra
writing. I need some shorthand. A debuging method to peer into the
program. Adding these things is not hard, but I can't let them
accumulate. First I will be verbose if needed, then add sugar. Keep it
small ...

Probably it is best if each piece of the program runs "clean", with
nothing piped to STDOUT or STDERR.