Add core/{R,sh}/common.loc
Typecasting is a major source of possible bugs in LOC. Especially when
dealing with dynamic languages, there are many different ways to
typecast a given function. For example, does `runif` have one positional
argument or three? It has one unnamed argument, but it has three that
are semantically necessary.

Questions:

 * How should I handle polymorphic functions?

 * How should I handle functions like `ls` which output data in a human
 readable form that needs parsing?

 * How opinionated should I be? Should I set default parameters?