Refactor and debug sh check code
Now anything typed as Bool in sh is recast automatically as 1 and 0, for
TRUE and FALSE, respectively. Bash functions can be written
idiomatically, e.g.

```
is_one () {
    test $1 -eq 1
}
```

All output is thrown away, and the exit code is translated into a
boolean. 1 for success, 0 for failure.

Now a function that takes a Bool argument may have the form:

```
 # foo :: Bool -> String
foo () {
    if test $1 -eq 0
    then
        echo FALSE
    else
        echo TRUE
    fi
}
```