Added very rough (untested) core Python functions
The need for library handling is becoming apparent. If a LOC script
includes multiple python sources, they will likely import a few of the
same packages. These will conflict. I need a system to track all
required libraries, unique them, and write them to the top of the
manifold pool.