Replace MULTI with '?' and NULL with void
Inside type and ontology sections, '?' now refers to 0 or more generic
types. This syntax matches that used in literature (see that gradual
type paper).

I replaced `NULL` with `void` as the IO type. This is a temporary type.
Eventually I will replace it with the Haskell IO. But IO in Haskell is a
constructor, which I do not yet support.