(2 fail) fix delimiter error in [*] joins
How should text be represented? I will leave this decision to the
specific language. In Python, I currently treat text as a list of
strings with whitespace stripped from the right hand side.

One thing, however, is not language-specific: Text type. Should it be
the same as `[String]`? Do I even need a `Text` type?

Practically, the difference between `String` and `Text` is that `String`
is passed as a literal argument, whereas `Text` is passed as a filename.
However, `[String]`, and all other vectors, are also passed as files.