Add basic type inference for star generics
I simply transfer the type for the input to the star, for example in

```
@type
foo :: * -> void
bar :: void -> A

@path
foo . bar
```

The type of foo will be inferred as `A -> void`.

More inferences are possible, for example inferring the output type
based on a calling manifold. However, here I implement only the simple
case. I will need to sit down at some point with a pencil and work
through all the possible inferences. Or maybe I can read up on type
theory.

Star-generics are a temporary feature of LOC. Eventually I will replace
them altogether with true generics.