Bug fix in star inference
Star generics are always atomic types (though they may be replaced with
non-atomics). I now check for this (not doing so led to SEGFAULTs).

I am not handling positionals very well. I could infer the type of at
least `Int`, `Num`, and `String`. For example in:

```
@path
foo . 1 1.234 "hello"
```

Though `Int` is a bit tricky, since it could just be a special case of
`Num`.