Refactor type handling, allow tuples
Now for a signature like

```
A :: a -> (a -> b) -> ([c], [d])
```

A structure like this is formed.

```
T_TYPE
    FT_FUNCTION
        FT_ATOMIC: a
        FT_FUNCTION:
            FT_ATOMIC: a
            FT_ATOMIC: b
        FT_TUPLE:
            FT_ARRAY:
                FT_ATOMIC: c
            FT_ARRAY:
                FT_ATOMIC: d
```

The frontend now supports tuples, but the backends do not. Also I have
no tests written for tuples on either end.

The universal representation for a flat tuple of primitives should be a
TAB-delimited row. This allows tables to be built with signatures like:
`[(Int, Int, String)]`. For nested or non-primative tuples, I will need
a richer representation. Probably JSON, although XML and YAML are also
possible.