Muddled a little more with type caster
I need to unify my universals. I've been trying to have "normal" forms
of all output. An Int is output as '7', a tuple as '4 hi'. But this just
doesn't generalize well. I end up with reams of special handling code.

So let's do JSON for everything. Perhaps I can add an optional humanizer
for the very final output.

The JSON file is a wrapper, the universal container. This allows the
interesting possibility of redefining the universal constructors to
carry additional state between calls.