Add test for tuples
I don't know how to express tuples generally in Bash. It doesn't seem to
have a well-supported data structure for this. Usually such things are
simply not done in the language. I am tempted to simply not support
nested tuples.

LOC doesn't require every language has a means of converting every type.
The greatest example of this is with function types, which will seldom
be interconvertible.

I have also been overcomplicating my JSON output. I don't need to
include type information. A single value is actually perfectly valid
JSON. For example

```
1
"this is a number"
[1,2,3]
[1,[2,"boo"]]
```

There we have it, atomics, arrays, and tuples.

Now there is the matter of whether I should use YAML or JSON. For now I
suppose I will use JSON. It is, after all, a subset of YAML. But for all
the data I will be passing currently, JSON is more than sufficient.