(28/63 fail) Adapt expectations around JSON null
Previously, if a pipeline returned no output, I printed nothing.
However, an empty JSON output isn't empty, it is "null". And there is a
difference between getting nothing and "null" (the former meaning
something went far South). And as a general principle, I don't want to
add special handling for special cases, rather I should go with the
conventions. So, a pipeline that does nothing will print "null" to
STDOUT.