Reshoe the dead horse
My implementation is totally deep-fried. Unsalvageable.

My generative grammar system is incomprehensible. The templates became
finer grain as I added more languages, so find I don't know what effect
a given piece has on the whole. More documentation might help.  Already
altering the grammar is nearly impossible. The code it generates is also
bad. I create functions where they are not needed. To solve this
problem, I would have to replace my current algorithm with one that
follows the composition, from exported sink to wells, creating required
functions as it goes. This is the right way to do it.

Trying to integrate Bash into this strictly typed framework is madness.
Perhaps it is possible, but I am not even close to a working solution.

Foreign calls from within an open composition are broken. They compile
to wildly incorrect code.

My type coercion system is riddled with holes.

The conversions to the common representation work only when the JSON
parsers on both sides happen to convert the datastrutures in the same
way. I need to recurse across the type, parsing the data into JSON as I
go. That was my intention. Doing this in a single language would be
tricky, writing an algorithm to generate code in a given language to do
this is even harder. Then I have to find a way to expose the templates
in a general way, allowing new languages and types to be added.

I should restart from nothing. But I can't. Not yet. I have to write the
paper. So I'll tape up the holes in my cardboard ship, throw the cargo
overboard, and ride to the bottom.

More concretely, I will add a bunch of tests, none of which pass. Move
them all to the "known problems" class. Scream silently and write the
damn paper. After the paper is out, I'll start designing the new
implementation. I will start with a minimal version of LOC, one which
includes the trickiest pieces (derefs, types, all that). Do the hard
parts first. Write the universal representation code as an independent
entity. The languages can import it as a library. Maybe I can spin off a
few other pieces.

The path to a solution of a subset of the problem, is not necessarily a
segment of the path to the optimal full solution. I would solve easy
problems and then solve harder problems starting from the easy basis.
Then I would find that the infrastructure that was sufficient for the
simple problem was inadequate for the large problem. This leads to the
program toppling, requiring a major refactor. My development cycle then
becomes solve easy problem 1. Revel in my progress. Extend to harder
problem 2. Fall into despair. Start a major refactor. Lament the fall
from a "Golden Age" where everything worked. Succeed in the refactor.
Rejoice once again. Then extend to harder problem 3 ...

How can I do better? Foresight. I need to plan the whole system (vaguely
perhaps). Then disassemble the system into modules. Pull these out,
perhaps into their own repositories. For each of these, think about what
is hardest, then solve this problem.

So what are the modules of LOC? The frontend and backend. These are
linked solely by LIL. The frontend is completely independent of the
backend. Indeed, to emphasize this, I ought to extract it into its own
repo.

The frontend separates into a lexer, parser, type-handler, a LIL-maker.
I should break LIL into two forms. An initial form that includes all
information in the LOC program, but without inference and coercion. A
second LIL generated from the first, it includes all generated functions
and inferred types (perhaps including wrapper functions). It should
contain detailed instructions for building the final programs.

The backend separates into the generative grammar that reads the final
LIL and builds the manifold nexus and manifold pools; and the coercion
library. The pools import libraries for `L a <-> U a` conversions and `U
a -> Bool` checking. It needs to have a system for specifying new types.

I can make it work.